ci-requirements.txt
mypy==0.761
Red-DiscordBot
